requirements.txt
Pillow==10.1.0
openpyxl==3.1.2